amc v2.0, ems v3.85, mc4plus v3.88, mc2plus v3.67, amc2c v 3.0.0, amcbldc v 2.0.10
these binaries implement ICC between the amc and the amc2c,
so that we can remove some CAN cabling in the wrist of ergoCub.

The extra CAN cable can however be kept as it does not harm.

the other binaries (amcbldc, ems, mc4plus, mc2plus)  have a new release because
of changes in code shared with the other boards (amc, amc2c)

these files were compiled w/
- icub-firmware sha 85aeba4929c8bb1a92f7b4807078edf1bdbe7cc0
- icub-firmware-shared sha 06946189f369a8653bfb41a0fc81a57aa77efc9e